client/requirements.txt
APScheduler==2.1.2
CherryPy==3.2.2
PyYAML==3.10
Pykka==1.2.0
argparse==1.2.1
beautifulsoup4==4.2.1
facebook-sdk==0.4.0
feedparser==5.1.3
mock==1.0.1
numpy==1.8.0
python-dateutil==2.1
python-mpd==0.3.0
pytz==2013b
quantities==0.10.1
semantic==1.0.3
six==1.6.1
ws4py==0.3.2
requests==2.1.0
Skype4Py==1.0.35
facepy==1.0.1
#mutagen-1.23 installato manualmente
